+++ src/main/java/kitchenpos/product/presentation/ProductRestController.java
@@ -1,7 +1,7 @@
-package kitchenpos.ui;
+package kitchenpos.product.presentation;
 
-import kitchenpos.application.ProductService;
-import kitchenpos.domain.Product;
+import kitchenpos.product.application.ProductService;
+import kitchenpos.product.domain.Product;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
