+++ src/test/java/kitchenpos/product/application/ProductServiceTest.java
@@ -1,9 +1,11 @@
-package kitchenpos.application;
+package kitchenpos.product.application;
 
+import kitchenpos.application.FakePurgomalumClient;
+import kitchenpos.product.infra.InMemoryProductRepository;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuRepository;
-import kitchenpos.domain.Product;
-import kitchenpos.domain.ProductRepository;
+import kitchenpos.product.domain.Product;
+import kitchenpos.product.domain.ProductRepository;
 import kitchenpos.infra.PurgomalumClient;
 import kitchenpos.menu.infra.InMemoryMenuRepository;
 import org.junit.jupiter.api.BeforeEach;
