+++ src/test/java/kitchenpos/product/infra/InMemoryProductRepository.java
@@ -1,7 +1,7 @@
-package kitchenpos.application;
+package kitchenpos.product.infra;
 
-import kitchenpos.domain.Product;
-import kitchenpos.domain.ProductRepository;
+import kitchenpos.product.domain.Product;
+import kitchenpos.product.domain.ProductRepository;
 
 import java.util.ArrayList;
 import java.util.HashMap;
